refactor: move legacy section renderer
